@@ -1,1 +1,1 @@
-ruby-2.6.1
+ruby-2.7.0
